refactor: move conflict messages to a shared location
