+++ packages/salesforcedx-vscode-core/src/commands/forceSourceDeploySourcePath.ts
@@ -9,6 +9,7 @@ import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode';
 import { ComponentSet } from '@salesforce/source-deploy-retrieve';
 import * as vscode from 'vscode';
 import { channelService } from '../channels';
+import { getConflictMessagesFor } from '../conflict/messages';
 import { nls } from '../messages';
 import { notificationService } from '../notifications';
 import { SfdxProjectConfig } from '../sfdxProject';
@@ -82,36 +83,25 @@ export const forceSourceDeploySourcePaths = async (
     }
   }
 
-  const messages: ConflictDetectionMessages = {
-    warningMessageKey: 'conflict_detect_conflicts_during_deploy',
-    commandHint: inputs => {
-      const commands: string[] = [];
-      (inputs as string[]).forEach(input => {
-        commands.push(
-          new SfdxCommandBuilder()
-            .withArg('force:source:deploy')
-            .withFlag('--sourcepath', input)
-            .build()
-            .toString()
-        );
-      });
-      const hints = commands.join('\n  ');
-
-      return hints;
-    }
-  };
-
-  const commandlet = new SfdxCommandlet<string[]>(
-    new SfdxWorkspaceChecker(),
-    new LibraryPathsGatherer(uris),
-    new LibraryDeploySourcePathExecutor(),
-    new CompositePostconditionChecker(
-      new SourcePathChecker(),
-      new TimestampConflictChecker(false, messages)
-    )
+  const messages:
+    | ConflictDetectionMessages
+    | undefined = getConflictMessagesFor(
+    'force_source_deploy_with_sourcepath_beta'
   );
 
-  await commandlet.run();
+  if (messages) {
+    const commandlet = new SfdxCommandlet<string[]>(
+      new SfdxWorkspaceChecker(),
+      new LibraryPathsGatherer(uris),
+      new LibraryDeploySourcePathExecutor(),
+      new CompositePostconditionChecker(
+        new SourcePathChecker(),
+        new TimestampConflictChecker(false, messages)
+      )
+    );
+
+    await commandlet.run();
+  }
 };
 
 export const getUriFromActiveEditor = (): vscode.Uri | undefined => {
